# libpci: Add PCIe ECAM access method