--- a/lib/internal.h
+++ b/lib/internal.h
@@ -136,5 +136,5 @@ void pci_free_caps(struct pci_dev *);
 extern struct pci_methods pm_intel_conf1, pm_intel_conf2, pm_linux_proc,
 	pm_fbsd_device, pm_aix_device, pm_nbsd_libpci, pm_obsd_device,
 	pm_dump, pm_linux_sysfs, pm_darwin, pm_sylixos_device, pm_hurd,
-	pm_mmio_conf1, pm_mmio_conf1_ext,
+	pm_mmio_conf1, pm_mmio_conf1_ext, pm_ecam,
 	pm_win32_cfgmgr32, pm_win32_kldbg, pm_win32_sysdbg;