… lib/ecam.c